+++ example/001-simple/indexOf/src/indexOf.ts
@@ -1,0 +1,6 @@
+/**
+ * Returns the position of the first occurrence of `pattern` in `text`
+ */
+export const indexOf = (text: string, pattern: string): number => {
+  return text.indexOf(pattern);
+};
