+++ example/004-stateMachine/musicPlayer/model-based/AddTrackCommand.ts
@@ -1,0 +1,21 @@
+import { MusicPlayerCommand, MusicPlayerModel } from './MusicPlayerModel';
+import { MusicPlayer } from '../src/MusicPlayer';
+import * as assert from 'assert';
+
+export class AddTrackCommand implements MusicPlayerCommand {
+  constructor(readonly position: number, readonly trackName: string) {}
+  check(m: MusicPlayerModel) {
+    return !m.tracksAlreadySeen[this.trackName];
+  }
+  run(m: MusicPlayerModel, p: MusicPlayer) {
+    const trackBefore = p.currentTrackName();
+    p.addTrack(this.trackName, this.position % (m.numTracks + 1)); // old model
+    assert.equal(p.playing(), m.isPlaying);
+    assert.equal(p.currentTrackName(), trackBefore);
+    ++m.numTracks;
+    m.tracksAlreadySeen[this.trackName] = true;
+  }
+  toString() {
+    return `AddTrack(${this.position}, "${this.trackName}")`;
+  }
+}
