+++ example/004-stateMachine/musicPlayer/model-based/NextCommand.ts
@@ -1,0 +1,22 @@
+import { MusicPlayerCommand, MusicPlayerModel } from './MusicPlayerModel';
+import { MusicPlayer } from '../src/MusicPlayer';
+import * as assert from 'assert';
+
+export class NextCommand implements MusicPlayerCommand {
+  check(m: MusicPlayerModel) {
+    return true;
+  }
+  run(m: MusicPlayerModel, p: MusicPlayer) {
+    const trackBefore = p.currentTrackName();
+    p.next();
+    assert.equal(p.playing(), m.isPlaying);
+    if (m.numTracks === 1) {
+      assert.equal(p.currentTrackName(), trackBefore);
+    } else {
+      assert.notEqual(p.currentTrackName(), trackBefore);
+    }
+  }
+  toString() {
+    return 'Next';
+  }
+}
