--- example/binary-trees/models/BinaryTree.ts
@@ -1,5 +1,0 @@
-export interface Tree<T> {
-  value: T;
-  left: Tree<T> | null;
-  right: Tree<T> | null;
-}
