--- example/contains/contains.js
@@ -1,5 +1,0 @@
-const contains = function(pattern, text) {
-  return text.indexOf(pattern) !== -1;
-};
-
-module.exports = { contains };
