--- example/model-based-testing/MusicPlayer.ts
@@ -1,67 +1,0 @@
-export interface MusicPlayer {
-  playing(): boolean;
-  currentTrackName(): string | null;
-  play(): void;
-  pause(): void;
-  addTrack(trackName: string, position: number): void;
-  next(): void;
-  jump(position: number): void;
-}
-
-/**
- * Basic state machine
- * Example of how fast-check can be derived for model based testing
- */
-class MusicPlayerRaw implements MusicPlayer {
-  private tracks: string[];
-  private isPlaying: boolean;
-  private playingId: number;
-
-  constructor(tracks: string[], readonly buggy: boolean) {
-    this.tracks = [...tracks];
-    this.isPlaying = false;
-    this.playingId = 0;
-  }
-
-  playing(): boolean {
-    return this.isPlaying;
-  }
-
-  currentTrackName(): string | null {
-    return this.playingId != null && this.playingId < this.tracks.length ? this.tracks[this.playingId] : null;
-  }
-
-  play(): void {
-    this.isPlaying = true;
-  }
-  pause(): void {
-    this.isPlaying = false;
-  }
-
-  addTrack(trackName: string, position: number): void {
-    this.tracks = [...this.tracks.slice(0, position), trackName, ...this.tracks.slice(position)];
-    if (!this.buggy && this.playingId >= position) {
-      ++this.playingId;
-    }
-  }
-
-  next(): void {
-    if (++this.playingId === this.tracks.length) {
-      this.playingId = 0;
-    }
-  }
-  jump(position: number): void {
-    this.playingId = position;
-  }
-}
-
-export class MusicPlayerA extends MusicPlayerRaw {
-  constructor(tracks: string[]) {
-    super(tracks, true);
-  }
-}
-export class MusicPlayerB extends MusicPlayerRaw {
-  constructor(tracks: string[]) {
-    super(tracks, false);
-  }
-}
